requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
certifi==2021.10.8
charset-normalizer==2.0.12
Deprecated==1.2.13
Django==3.2
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-filter==21.1
django-parler==2.3
django-parler-rest==2.1
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
idna==3.3
packaging==21.3
Pillow==9.0.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2022.1
redis==4.2.2
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
wrapt==1.14.0
